feat(mgmt): reorganize code
Move config database outside of model definition submodule.
