+++ mgmt/src/frr/renderer/mod.rs
@@ -14,7 +14,7 @@ pub mod statics;
 pub mod vrf;
 
 use crate::frr::renderer::builder::{ConfigBuilder, Render};
-use crate::models::external::configdb::gwconfig::GwConfig;
+use crate::models::external::gwconfig::GwConfig;
 use crate::models::internal::InternalConfig;
 
 fn render_metadata(config: &GwConfig) -> String {
