--- mgmt/src/models/external/configdb/mod.rs
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Open Network Fabric Authors
-
-pub mod gwconfig;
-pub mod gwconfigdb;
