+++ mgmt/src/processor/mod.rs
@@ -5,6 +5,7 @@
 //! This module implements the core logic to determine and build internal configurations.
 
 pub mod confbuild;
+pub mod gwconfigdb;
 mod namegen;
 pub mod proc;
 pub mod tests;
